requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
aiofiles==23.2.1
redis==5.0.1
celery==5.3.4
pytz==2023.3
gTTS==2.3.2
pydub==0.25.1
ffmpeg-python==0.2.0
aiohttp==3.9.1
# Заменяем asyncpg на чистый SQLAlchemy
# asyncpg не нужен, убираем его